Send a GA event when user requests CM datasheet
When the user triggers submission of the form on the
CM page send an event to GA. This allows for it to be
configured as a goal in GA to see how well we are
doing at conversion.

This should perhaps be on all the forms, but the global
way it is triggered makes that a bit tricky. There may
be a way to examine the data to see which form it is
to include that in the event, but I haven't looked.