Do not specify namespace for DO DNS01 secret
The examples for `Issuer` do not specify a namespace, so when we copy/paste the examples we stick them  in the current namespace. Specifying that this secret goes in the `cert-manager` namespace potentially puts it into a different namespace than the `cert-manager` pods, making it a lot more confusing when it doesn't work.

Signed-off-by: Jamie Gaskins <jgaskins@hey.com>