Adjust list of Subject Alternative Names
A description of resource `Certificate` has `foo.example.com` and `bar.example.com` in the `dnsNames` section, which is different from the diagram that has `example.com` and `foo.example.com` domains. When I was looking at the schema in the article, I thought that the part related to `bar.example.com` is to follow, but it didn't and therefore was a bit confusing.