remove alex check
we never looked at the output from this command - it didn't fail the
build. it also has so many false positives that it would take a
significant amount of effort to be able to use it effectively.

I like the idea of tools like this, but I don't think this specific tool
- with the false-positives - is providing any value to us here.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>