remove old .gitignores
previously we'd not checked api-docs into the repo for docs and
next-docs, relying on generating them at deploy time

as part of the refresh I checked them in; we still re-generate them at
deploy time in case the checked-in version is out of date, but it's
better IMO to have a maybe-a-little-out-of-date version checked in than
to have nothing checked in.

in any case, these gitignores are no longer needed, and also just don't
seem to be doing anything!

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>