Various updates to nginx ingress tutorial
This fixes a bunch of language, tweaks links, and generally makes
improvements. The tutorial isn't finished by any means, but this feels
like a bunch of incremental improvements.

Also changes links on the tutorial from cert-manager.io to github user
content - the links were broken by the refresh, and this patches over
that (and might actually be an improvement, anyway)

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>