change opening on contributing guide, remove second sidebar
I don't think it's worth linking to everything a second time. People can
find all of those links in the sidebar.

If we need to call something out specifically, we can do that in the
README, but I don't see the value of listing all the links a second
time.

Also changes the language of the opening section and makes minor tweaks

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>