document the newer simpler versioning and the old release processes
We didn't have any documentation in this process on how to release older
versions of cert-manager. That process has a couple of gotchas,
especially for 1.5 and below. This commit adds some pointers for that
older release process.

It also updates the process now to reflect that the release version can
be used to refer to staged resources for versions v1.8.1+ and v1.9+,
removing a step from the process.

Signed-off-by: Ashley Davis <ashley.davis@jetstack.io>