Modifications after trying the istio-csr tutorial
This commit makes two small tweaks to the istio-csr tutorial on the website.

The first is that the final clean up command will fail, following the instructions verbatim
will produce an error, so it has been updated to just `kind delete cluster`, which successfully
deletes the kind cluster. 

```
$ kind delete cluster kind
ERROR: unknown command "kind" for "kind delete cluster"
```

The second change is to add that the `isitoctl install` command will prompt for user input
to confirm the choice of profile. The user should confirm the choice. 

Signed-off-by: David Bond <davidsbond@users.noreply.github.com>