fix the open graph tag "og:title"
The og:title is important because that's the first thing
people see when they share a cert-maanger.io link in Twitter
or on Slack. Today, all the pages have the same og:title:

  cert-manager - Documentation

The title of the page that we set in each frontmatter is
discarded. This change fixes that.

Signed-off-by: Maël Valais <mael@vls.dev>