release-process: create the release-next periodics on beta
Our release instructions were wrong to suggest we should be running the
periodic tests for the "next" release branch after the "initial alpha".
The results of these tests are equal to the periodics we have on master
until master diverges from the "next" release branch.

From now on, let's only create the periodic after we have cut the first
beta, which will be the moment where master and the release-* branch
will start to diverge.

Signed-off-by: Maël Valais <mael@vls.dev>