post-release 1.12: update the release process with the latest process
I re-arranged the website PR instructions so that the pre-release steps for the
website are shown first. I also made it clearer that we produce two types of
release notes: the GitHub release description and the "Release Note" page on the
website.

- Mention "How we determine supported Kubernetes".
- Mention how to transform GitHub references into links in the
  `release-notes.md` file.
- Mention bump 1.12 in generate-new-import-path-docs
- Mention the freeze of docs/ that gets copied into v1.12-docs/.
- Mention how to bump the gen scripts.
- Show a command that to bump versions in install instructions.
- Mention manifest.json that needs to be updated.
- Mention the file release-notes/README.md's index that needs updating.

I also had to re-order the bullet points due to conflicts with Irbe's changes to
the release process document. Apologies for the noise that these changes
introduce!

Signed-off-by: Maël Valais <mael@vls.dev>