update csi-driver project page
This page has a lot of prose which I think benefits from clarification.
There are also many typos and the first section of our installation
instructions talks about an ancient Kubernetes version, with
instructions which will apply to approximately nobody.

This PR should hopefully make the docs simpler and easier to follow.

Signed-off-by: Ashley Davis <ashley.davis@venafi.com>