releases: Move v1.14 to "Currently supported"
Release 1.14 has been officially tagged.

Tags v1.14.0 and v1.14.1 (hotfix) were released on Feb 2, 2024. The releases document is updated to reflect this, and the provisional End of Life date is marked as four months in the future, falling on the nearest working day to allow for a last minute release without risk of breaking someone's weekend.

Signed-off-by: Valters Jansons <sigv@users.noreply.github.com>