Remove $ from bash code snippets
While $ is a good way to differentiate between a command and its output, it is best to leave it out whenever there is no output. This way, the user can copy the command by clicking the copy button on the rendered website without having to worry about the leading $ when pasting the command.

This was already the case for some snippets. This change removes the leading $ for some more snippets.

Signed-off-by: Andreas Lindhé <7773090+lindhe@users.noreply.github.com>